refactor: validations and constants
