+++ src/constants/dto-validations.ts
@@ -1,0 +1,4 @@
+export const DTO_VALIDATIONS = {
+  PROFESSIONAL_NAME_LENGTH: 24,
+  PROFESSIONAL_ABOUT_LENGTH: 512,
+};
