--- src/validators/index.ts
@@ -1,1 +1,0 @@
-export * from './is-date-after';
